Separate character endpoint(s) into its own router
